requirements.txt
--index-url https://pypi.org/simple
python-telegram-bot==20.8
qrcode==7.4.2
pillow==11.0.0
pymongo==4.6.1
python-dotenv==1.0.1
dnspython==2.6.1 